requirements.txt
# Web framework
Flask==2.3.4

# Computer vision / face recognition
opencv-python==4.8.1.78
insightface==0.8.5
numpy==1.25.2
scikit-learn==1.3.2

# Email sending
secure-smtplib==0.1  # optional, usually smtplib comes with Python

# Data handling
pandas==2.1.1  # if you handle CSVs
python-dotenv==1.0.1  # for environment variables like secret keys

# Optional for MongoDB
pymongo==4.5.0
dnspython==2.3.0  # required for some MongoDB URIs

# Optional for image handling / emails
Pillow==10.1.0

# For file uploads
Werkzeug==2.3.7
